\name{BayesLCA - Utilities}
\alias{weighted.var}
\alias{Zsamp}
\alias{xlogy}
\alias{Zscore.internal}
\alias{MAP}
\alias{unMAP}

\title{
Internal BayesLCA function
}
\description{
An internal function not intended to be directly called by users.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
